Fix Docker Compose ftdetect
This upgrade simplifies the Docker Compose file type detection down to
one regex. It also prevents that ftdetect from detecting itself.
Finally, it adds support for .yaml extensions and drops support for
docker-compose files without an extension, as that is madness.

I also changed all the `setf`s to `set ft` because it wasn't working in
Neovim for some reason.